Refactor ICheepService to separate interfaces to ICheeps and IAuthor + fix firefox ui searchbar
