+++ src/Web/Interfaces/ICheepService.cs
@@ -4,14 +4,22 @@ namespace Web.Interfaces;
 
 public interface ICheepService
 {
-    public Task<IEnumerable<Author>> SearchAuthors(string searchQuery, int page);
-    public Task<IEnumerable<Cheep>> SearchCheeps(string searchQuery, int page);
+    public interface ICheeps
+    {
+        public List<Cheep> GetCheeps();
 
-    public Task<IEnumerable<Cheep>> GetAllCheeps(int page);
+        public Task<IEnumerable<Cheep>> SearchCheeps(string searchQuery, int page);
 
+        public Task<IEnumerable<Cheep>> GetAllCheeps(int page);
 
-    public List<Cheep> GetCheepsFromAuthor(string author, int page);
-    public List<Cheep> GetCheeps();
-
-    public List<Author> GetAuthors();
+        public void StoreCheep(Cheep cheep);
+    }
+    public interface IAuthor
+    {
+        public List<Author> GetAuthors();
+        public Task<IEnumerable<Author>> SearchAuthors(string searchQuery, int page);
+        public Task<IEnumerable<Cheep>> GetCheepsFromAuthor(string author, int page);
+    }
 }
+
+
