+++ src/Web/Program.cs
@@ -10,7 +10,9 @@ builder.Services.AddDbContext<CheepDbContext>(options => options.UseSqlite(conne
 
 builder.Services.AddRazorPages();
 
-builder.Services.AddScoped<ICheepService, CheepService>();
+builder.Services.AddScoped<ICheepService.ICheeps, CheepServiceCheeps>();
+builder.Services.AddScoped<ICheepService.IAuthor, CheepServiceAuthor>();
+
 
 var app = builder.Build();
